Move resource into proper folder
