+++ prime-angular-frontend/src/app/modules/paper-enrolment/shared/services/paper-enrolment-resource.service.spec.ts
